Move condition logic into its own file
