+++ src/menu/menuItem.ts
@@ -1,11 +1,7 @@
 import { QuickPickItem } from "vscode";
-import { ActionType, BindingItem, OverrideBindingItem } from "../bindingItem";
+import { ActionType, BindingItem } from "../bindingItem";
+import { Condition, evalCondition, getCondition, isConditionKeyEqual } from "../condition";
 import { specializeBindingKey } from "../utils";
-
-interface Condition {
-    when?: string,
-    languageId?: string,
-}
 
 export interface MenuSelectionResult {
     items?: BaseMenuItem[],
@@ -95,56 +91,6 @@ export class CommandsMenuItem extends BaseMenuItem {
     }
 }
 
-function evalCondition(stored?: Condition, evaluatee?: Condition) {
-    if (evaluatee && stored) {
-        let result = true;
-        if (stored.when) {
-            result = result && (stored.when === evaluatee.when);
-        }
-        if (stored.languageId) {
-            result = result && (stored.languageId === evaluatee.languageId);
-        }
-        return result;
-    }
-    // For if they are both undefined or null
-    return stored === evaluatee;
-}
-
-function isConditionEqual(condition1?: Condition, condition2?: Condition) {
-    if (condition1 && condition2) {
-        let result = true;
-        result = result && (condition1.when === condition2.when);
-        result = result && (condition1.languageId === condition2.languageId);
-        return result;
-    }
-    // For if they are both undefined or null
-    return condition1 === condition2;
-}
-
-function isConditionKeyEqual(key1?: string, key2?: string) {
-    return isConditionEqual(getCondition(key1), getCondition(key2));
-}
-
-function getCondition(key?: string): Condition | undefined {
-    if (key && key.length > 0) {
-        const props = key.split(";");
-        const r = props.reduce((result, prop) => {
-            const [key, value] = prop.split(":");
-            result[key] = value;
-            return result;
-        }, {} as Record<string, string>);
-
-        // Check to make sure at least the one property so we don't create
-        // { when: undefined, languagedId: undefined }
-        if ("when" in r || "languageId" in r) {
-            return {
-                when: r["when"],
-                languageId: r["languageId"]
-            };
-        }
-    }
-    return undefined;
-}
 
 class ConditionalsMenuItem extends BaseCollectionMenuItem {
     constructor(item: BindingItem) {
